refactor: ItemService를 Query(읽기전용), Command(쓰기전용) 으로 분리
